requirements.txt
streamlit
Pillow
numpy
torch
transformers
timm